requirements.txt
prefect==2.14.18
pyspark==3.5.0
requests==2.31.0
python-dotenv==1.0.1
minio==7.2.4
griffe==0.39.1
pytest==7.4.4
pandas==2.1.4
